+++ src/main/java/src/shared/FXMLFilenames.java
@@ -6,7 +6,8 @@ public enum FXMLFilenames {
     SET_TOTAL_FUNDS_POPUP("set_total_funds_popup"),
     CREATE_PAY_PERIOD_POPUP("create_pay_period_popup"),
     CREATE_EXPENSE_POPUP("create_expense_popup"),
-    UPDATE_EXPENSE_POPUP("update_expense_popup");
+    UPDATE_EXPENSE_POPUP("update_expense_popup"),
+    ADD_AMOUNT_SPENT_POPUP("add_amount_spent_popup");
     
 
     private final String fxmlFilename;