+++ CHANGELOG.md
@@ -2,3 +2,14 @@
 - Auto-expand TableView column width based on cell values
 - Fix bug when closing Pay Period Frequency Selection popup without clicking the submit button
 - Created an input option to set the current amount spent to the spending limit value for a newly-created Expense
+
+## Version 1.1.0
+- Update openjfx to version 25-ea+29
+- Update row color based on Expense values
+- Edit an Expense when double-clicking on a TableView row
+- Update when the "Already Paid" checkbox is interactable
+- Remove the "Expense Type" column from the TableViews
+- Update spending limit of expenses in a previous pay period when creating a new pay period
+- Fix bug where updating pay period frequency prior to having a pay period caused an error
+- Update expense page button icons
+- Add feature to allow values to be added to the Current Amount Spent of an Expense